+++ skyfield/tests/test_io.py
@@ -4,7 +4,6 @@ import os
 import shutil
 import tempfile
 from contextlib import contextmanager
-from datetime import date
 try:
     from unittest.mock import patch
 except ImportError:
@@ -35,13 +34,10 @@ def file_contents(load):
         return f.read()
 
 @contextmanager
-def on(load, year, month, day):
-    fake_date = date(year, month, day)
+def fake_download(load):
     download = lambda *args, **kw: save_file(load, new_content)
     with patch('skyfield.iokit.download', download):
-        # Python 2.6 does not support the comma "with" statement, so:
-        with patch('skyfield.iokit.today', lambda *args: fake_date):
-            yield
+        yield
 
 # The tests.
 
@@ -65,24 +61,11 @@ def test_open_in_subdirectory(load):
     assert data == b'example text\n'
 
 def test_missing_file_gets_downloaded(load):
-    with on(load, 2016, 1, 15):
+    with fake_download(load):
         data = load('deltat.data')
-        assert file_contents(load).endswith(b' 68.1577\n')
-    assert data[1][-1] == 68.1577
-
-def test_11_month_old_file_gets_reused(load):
-    save_file(load)
-    with on(load, 2016, 12, 15):
-        data = load('deltat.data')
-        assert file_contents(load).endswith(b' 68.1024\n')
-    assert data[1][-1] == 68.1024
-
-def test_12_month_old_file_gets_redownloaded(load):
-    save_file(load)
-    with on(load, 2017, 1, 15):
-        data = load('deltat.data')
-        assert file_contents(load).endswith(b' 68.1577\n')
-    assert data[1][-1] == 68.1577
+    print(repr(file_contents(load)[:-20]))
+    assert file_contents(load).endswith(b' 68.1577\n')
+    assert data[-1] == 68.1577
 
 def test_builtin_timescale(load):
     ts = load.timescale()